requirements.txt
discord.py==2.0.1
facebook_scraper==0.2.59
